requirements.txt
filelock==3.20.0
fsspec==2025.12.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
setuptools==80.9.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
